refactor(simple): move generate_tasks/agents from task/agent into sim
- agent.py and task.py now only define Agent and Task classes
- generate_tasks() and generate_agents() relocated to sim.py
- sim.py imports Task/Agent classes directly and owns instantiation logic
