+++ scenarios/simple/sim/sim.py
@@ -1,8 +1,41 @@
 from modules.base_sim import BaseSim
-from modules.utils import ResultSaver
-from scenarios.simple.sim.task import generate_tasks
-from scenarios.simple.sim.agent import generate_agents
-import pygame
+from modules.utils import ResultSaver, config, generate_positions
+from scenarios.simple.sim.task import Task
+from scenarios.simple.sim.agent import Agent
+
+
+def generate_tasks(task_quantity=None, task_id_start=0, seed=None):
+    if task_quantity is None:
+        task_quantity = config['tasks']['quantity']
+    task_locations = config['tasks']['locations']
+
+    tasks_positions = generate_positions(task_quantity,
+                                        task_locations['x_min'],
+                                        task_locations['x_max'],
+                                        task_locations['y_min'],
+                                        task_locations['y_max'],
+                                        radius=task_locations['non_overlap_radius'],
+                                        seed=seed)
+
+    tasks = [Task(idx + task_id_start, pos) for idx, pos in enumerate(tasks_positions)]
+    return tasks
+
+
+def generate_agents(tasks_info, seed=None):
+    agent_quantity = config['agents']['quantity']
+    agent_locations = config['agents']['locations']
+
+    agents_positions = generate_positions(agent_quantity,
+                                          agent_locations['x_min'],
+                                          agent_locations['x_max'],
+                                          agent_locations['y_min'],
+                                          agent_locations['y_max'],
+                                          radius=agent_locations['non_overlap_radius'],
+                                          seed=seed)
+
+    agents = [Agent(idx, pos, tasks_info) for idx, pos in enumerate(agents_positions)]
+    return agents
+
 
 class Sim(BaseSim):
     def __init__(self, config):
@@ -29,30 +62,7 @@ class Sim(BaseSim):
 
     def save_results(self):
         pass
-        # # Save gif
-        # if self.save_gif and self.rendering_mode == "Screen":        
-        #     self.recording = False
-        #     print("Recording stopped.")
-        #     self.result_saver.save_gif(self.frames)          
-     
 
-        # # Save time series data
-        # if self.save_timewise_result_csv:        
-        #     csv_file_path = self.result_saver.save_to_csv("timewise", self.data_records, ['time', 'agents_total_distance_moved', 'agents_total_task_amount_done', 'remaining_tasks', 'tasks_total_amount_left'])          
-        #     self.result_saver.plot_timewise_result(csv_file_path)
-        
-        # # Save agent-wise data            
-        # if self.save_agentwise_result_csv:        
-        #     variables_to_save = ['agent_id', 'task_amount_done', 'distance_moved']
-        #     agentwise_results = self.result_saver.get_agentwise_results(self.agents, variables_to_save)                        
-        #     csv_file_path = self.result_saver.save_to_csv('agentwise', agentwise_results, variables_to_save)
-            
-        #     self.result_saver.plot_boxplot(csv_file_path, variables_to_save[1:])
-
-        # # Save yaml
-        # if self.save_config_yaml:                
-        #     self.result_saver.save_config_yaml()           
-   
     def record_timewise_result(self):
         agents_total_distance_moved = sum(agent.distance_moved for agent in self.agents)
         agents_total_task_amount_done = sum(agent.task_amount_done for agent in self.agents)
