+++ scenarios/simple/sim/task.py
@@ -1,6 +1,6 @@
 import pygame
 import random
-from modules.utils import config, generate_positions, generate_task_colors
+from modules.utils import config, generate_task_colors
 dynamic_task_generation = config['tasks'].get('dynamic_task_generation', {})
 max_generations = dynamic_task_generation.get('max_generations', 0) if dynamic_task_generation.get('enabled', False) else 0
 tasks_per_generation = dynamic_task_generation.get('tasks_per_generation', 0) if dynamic_task_generation.get('enabled', False) else 0
@@ -29,19 +29,3 @@ class Task(BaseTask):
             screen.blit(text_surface, (self.position[0], self.position[1]))
 
 
-def generate_tasks(task_quantity=None, task_id_start = 0, seed=None):
-    if task_quantity is None:
-        task_quantity = config['tasks']['quantity']        
-    task_locations = config['tasks']['locations']
-
-    tasks_positions = generate_positions(task_quantity,
-                                        task_locations['x_min'],
-                                        task_locations['x_max'],
-                                        task_locations['y_min'],
-                                        task_locations['y_max'],
-                                        radius=task_locations['non_overlap_radius'],
-                                        seed=seed)
-
-    # Initialize tasks
-    tasks = [Task(idx + task_id_start, pos) for idx, pos in enumerate(tasks_positions)]
-    return tasks
